snapshot: further simplify the rbd source
