+++ src/backy/rbd/chunked/__init__.py
@@ -1,0 +1,26 @@
+from typing import TypeAlias
+
+from .chunk import Chunk
+from .file import File
+from .store import Store
+
+__all__ = [
+    "Chunk",
+    "File",
+    "Store",
+    "Hash",
+    "BackendException",
+    "InconsistentHash",
+]
+
+Hash: TypeAlias = str
+
+
+class BackendException(IOError):
+    pass
+
+
+class InconsistentHash(BackendException):
+    def __init__(self, expected, actual):
+        self.expected = expected
+        self.actual = actual
