+++ src/backy/rbd/source.py
@@ -1,37 +1,34 @@
+import json
 import os
 import subprocess
 import time
+import uuid
 from enum import Enum
 from pathlib import Path
 from typing import IO, Any, Optional, Set
 
+import consulate
 from structlog.stdlib import BoundLogger
 
 import backy
+import backy.utils
+from backy.ext_deps import BACKY_EXTRACT
+from backy.repository import Repository
+from backy.revision import Revision, Trust
+from backy.source import Source
+from backy.utils import (
+    CHUNK_SIZE,
+    END,
+    TimeOut,
+    TimeOutError,
+    copy,
+    posix_fadvise,
+    report_status,
+)
 
-from ..ext_deps import BACKY_EXTRACT
-from ..repository import Repository
-from ..revision import Revision, Trust
-from ..source import Source
-from ..utils import CHUNK_SIZE, END, copy, posix_fadvise, report_status
-from .chunked.chunk import BackendException, Chunk, Hash
-from .chunked.file import File
-from .chunked.store import Store
-from .quarantine import QuarantineStore
-from .sources.flyingcircus.source import FlyingCircusRootDisk
-
-# Locking strategy:
-#
-# - You can only run one backup of a machine at a time, as the backup will
-#   interact with this machines' list of snapshots and will get confused
-#   if run in parallel.
-# - You can restore while a backup is running.
-# - You can only purge while nothing else is happening.
-# - Trying to get a shared lock (specifically purge) will block and wait
-#   whereas trying to get an exclusive lock (running backups, purging) will
-#   immediately give up.
-# - Locking is not re-entrant. It's forbidden and protected to call another
-#   locking main function.
+from .chunked import BackendException, Chunk, File, Hash, Store
+from .quarantine import QuarantineReport, QuarantineStore
+from .rbd import RBDClient
 
 
 class RestoreBackend(Enum):
@@ -43,16 +40,21 @@ class RestoreBackend(Enum):
         return self.value
 
 
-class RbdSource(Source):
+class RestoreArgs:
+    target: str
+    backend: RestoreBackend
+
+
+class RbdSource(Source[RestoreArgs]):
     type_ = "rbd"
     subcommand = "backy-rbd"
 
-    source: FlyingCircusRootDisk
+    source: "FlyingCircusRootDisk"
     store: Store
     quarantine: QuarantineStore
     log: BoundLogger
 
-    def __init__(self, rbdsource: FlyingCircusRootDisk, log: BoundLogger):
+    def __init__(self, rbdsource: "FlyingCircusRootDisk", log: BoundLogger):
         self.source = rbdsource
         self.log = log.bind(subsystem="rbdsource")
 
@@ -65,7 +67,7 @@ class RbdSource(Source):
         super().bind(repository)
         # TODO: move quarantine to repo
         self.quarantine = QuarantineStore(repository.path, self.log)
-        self.store = Store(repository.path / "chunks")
+        self.store = Store(repository.path / "chunks", self.log)
 
     def _path_for_revision(self, revision: Revision) -> Path:
         return self.repository.path / revision.uuid
@@ -227,14 +229,9 @@ class RbdSource(Source):
 
     # This needs no locking as it's only a wrapper for restore_file and
     # restore_stdout and locking isn't re-entrant.
-    def restore(
-        self,
-        revision: Revision,
-        target: str,
-        restore_backend: RestoreBackend = RestoreBackend.AUTO,
-    ) -> None:
-        s = ChunkedFileBackend(revision, self.log).open("rb")
-        if restore_backend == RestoreBackend.AUTO:
+    def restore(self, revision: Revision, args: RestoreArgs) -> None:
+        s = self.open(revision)
+        if args.backend == RestoreBackend.AUTO:
             if self.backy_extract_supported(s):
                 restore_backend = RestoreBackend.RUST
             else:
@@ -242,12 +239,12 @@ class RbdSource(Source):
             self.log.info("restore-backend", backend=restore_backend.value)
         if restore_backend == RestoreBackend.PYTHON:
             with s as source:
-                if target != "-":
-                    self.restore_file(source, target)
+                if args.target != "-":
+                    self.restore_file(source, args.target)
                 else:
                     self.restore_stdout(source)
         elif restore_backend == RestoreBackend.RUST:
-            self.restore_backy_extract(revision, target)
+            self.restore_backy_extract(revision, args.target)
 
     def backy_extract_supported(self, file: "backy.rbd.chunked.File") -> bool:
         log = self.log.bind(subsystem="backy-extract")
@@ -263,7 +260,7 @@ class RbdSource(Source):
             if not version.startswith("backy-extract"):
                 log.debug("unknown-version")
                 return False
-        except:
+        except Exception:
             log.debug("unavailable")
             return False
         return True
@@ -282,7 +279,8 @@ class RbdSource(Source):
         )
         if return_code:
             raise RuntimeError(
-                f"backy-extract failed with return code {return_code}. Maybe try `--backend python`?"
+                f"backy-extract failed with return code {return_code}. "
+                "Maybe try `--backend python`?"
             )
 
     @Repository.locked(target=".purge", mode="shared")
@@ -311,3 +309,226 @@ class RbdSource(Source):
                 if not chunk:
                     break
                 target.write(chunk)
+
+
+class CephRBD:
+    """The Ceph RBD source.
+
+    Manages snapshots corresponding to revisions and provides a verification
+    that tries to balance reliability and performance.
+    """
+
+    pool: str
+    image: str
+    always_full: bool
+    log: BoundLogger
+    rbd: RBDClient
+    revision: Revision
+    snapshot_timeout = 90
+
+    def __init__(self, config: dict, log: BoundLogger):
+        self.pool = config["pool"]
+        self.image = config["image"]
+        self.always_full = config.get("full-always", False)
+        self.log = log.bind(subsystem="ceph")
+        self.rbd = RBDClient(self.log)
+        self.vm = config["vm"]
+        self.consul_acl_token = config.get("consul_acl_token")
+
+    def ready(self) -> bool:
+        """Check whether the source can be backed up.
+
+        For RBD sources this means the volume exists and is accessible.
+
+        """
+        try:
+            if self.rbd.exists(self._image_name):
+                return True
+        except Exception:
+            self.log.exception("not-ready")
+        return False
+
+    def __call__(self, revision):
+        self.revision = revision
+        return self
+
+    def __enter__(self):
+        snapname = "backy-{}".format(self.revision.uuid)
+        self.create_snapshot(snapname)
+        return self
+
+    def create_snapshot(self, snapname: str) -> None:
+        """An overridable method to allow different ways of creating the
+        snapshot.
+        """
+        self.rbd.snap_create(self._image_name + "@" + snapname)
+
+    @property
+    def _image_name(self) -> str:
+        return "{}/{}".format(self.pool, self.image)
+
+    def __exit__(self, exc_type=None, exc_val=None, exc_tb=None):
+        self._delete_old_snapshots()
+
+    def get_parent(self) -> Optional[Revision]:
+        if self.always_full:
+            self.log.info("backup-always-full")
+            return None
+        revision = self.revision
+        while True:
+            parent = revision.get_parent()
+            if not parent:
+                self.log.info("backup-no-valid-parent")
+                return None
+            if not self.rbd.exists(self._image_name + "@backy-" + parent.uuid):
+                self.log.info(
+                    "ignoring-rev-without-snapshot",
+                    revision_uuid=parent.uuid,
+                )
+                revision = parent
+                continue
+            # Ok, it's trusted and we have a snapshot. Let's do a diff.
+            return parent
+
+    def diff(self, target: "RbdSource", parent: Revision) -> None:
+        self.log.info("diff")
+        snap_from = "backy-" + parent.uuid
+        snap_to = "backy-" + self.revision.uuid
+        s = self.rbd.export_diff(self._image_name + "@" + snap_to, snap_from)
+        with s as source, target.open(self.revision, parent) as target_:
+            bytes = source.integrate(target_, snap_from, snap_to)
+        self.log.info("diff-integration-finished")
+
+        self.revision.stats["bytes_written"] = bytes
+
+        # TMP Gather statistics to see where to optimize
+        from backy.rbd.chunked.chunk import chunk_stats
+
+        self.revision.stats["chunk_stats"] = chunk_stats
+
+    def full(self, target: File) -> None:
+        self.log.info("full")
+        s = self.rbd.export(
+            "{}/{}@backy-{}".format(self.pool, self.image, self.revision.uuid)
+        )
+        copied = 0
+        with s as source, target as target_:
+            while True:
+                buf = source.read(4 * backy.utils.MiB)
+                if not buf:
+                    break
+                target_.write(buf)
+                copied += len(buf)
+        self.revision.stats["bytes_written"] = copied
+
+        # TMP Gather statistics to see if we actually are aligned.
+        from backy.rbd.chunked.chunk import chunk_stats
+
+        self.revision.stats["chunk_stats"] = chunk_stats
+
+    def verify(self, target: File, quarantine: QuarantineStore) -> bool:
+        s = self.rbd.image_reader(
+            "{}/{}@backy-{}".format(self.pool, self.image, self.revision.uuid)
+        )
+        self.revision.stats["ceph-verification"] = "partial"
+
+        with s as source, target as target_:
+            self.log.info("verify")
+            return backy.utils.files_are_roughly_equal(
+                source,
+                target_,
+                report=lambda s, t, o: quarantine.add_report(
+                    QuarantineReport(s, t, o)
+                ),
+            )
+
+    def _delete_old_snapshots(self) -> None:
+        # Clean up all snapshots except the one for the most recent valid
+        # revision.
+        # Previously we used to remove all snapshots but the one for this
+        # revision - which is wrong: broken new revisions would always cause
+        # full backups instead of new deltas based on the most recent valid
+        # one.
+        # XXX this will break if multiple servers are active
+        if not self.always_full and self.revision.repository.local_history:
+            keep_snapshot_revision = self.revision.repository.local_history[
+                -1
+            ].uuid
+        else:
+            keep_snapshot_revision = None
+        for snapshot in self.rbd.snap_ls(self._image_name):
+            if not snapshot["name"].startswith("backy-"):
+                # Do not touch non-backy snapshots
+                continue
+            uuid = snapshot["name"].replace("backy-", "")
+            if uuid != keep_snapshot_revision:
+                time.sleep(3)  # avoid race condition while unmapping
+                self.log.info(
+                    "delete-old-snapshot", snapshot_name=snapshot["name"]
+                )
+                try:
+                    self.rbd.snap_rm(self._image_name + "@" + snapshot["name"])
+                except Exception:
+                    self.log.exception(
+                        "delete-old-snapshot-failed",
+                        snapshot_name=snapshot["name"],
+                    )
+
+    def create_snapshot(self, name: str) -> None:
+        consul = consulate.Consul(token=self.consul_acl_token)
+        snapshot_key = "snapshot/{}".format(str(uuid.uuid4()))
+        self.log.info(
+            "creating-snapshot",
+            snapshot_name=name,
+            snapshot_key=snapshot_key,
+        )
+
+        consul.kv[snapshot_key] = {"vm": self.vm, "snapshot": name}
+
+        time.sleep(3)
+        try:
+            timeout = TimeOut(
+                self.snapshot_timeout, interval=2, raise_on_timeout=True
+            )
+            while timeout.tick():
+                for snapshot in self.rbd.snap_ls(self._image_name):
+                    if snapshot["name"] == name:
+                        return
+        except TimeOutError:
+            # The VM might have been shut down. Try doing a regular Ceph
+            # snapshot locally.
+            super(FlyingCircusRootDisk, self).create_snapshot(name)
+        except KeyboardInterrupt:
+            raise
+        finally:
+            # In case the snapshot still gets created: the general snapshot
+            # deletion code in ceph/source will clean up unused backy snapshots
+            # anyway. However, we need to work a little harder to delete old
+            # snapshot requests, otherwise we've sometimes seen those not
+            # getting deleted and then re-created all the time.
+            for key in list(consul.kv.find("snapshot/")):
+                try:
+                    s = consul.kv[key]
+                except KeyError:
+                    continue
+                try:
+                    s = json.loads(s)
+                except json.decoder.JSONDecodeError:
+                    # Clean up garbage.
+                    self.log.warning(
+                        "create-snapshot-removing-garbage-request",
+                        snapshot_key=key,
+                    )
+                    del consul.kv[key]
+                if s["vm"] != self.vm:
+                    continue
+                # The knowledge about the `backy-` prefix  isn't properly
+                # encapsulated here.
+                if s["snapshot"].startswith("backy-"):
+                    self.log.info(
+                        "create-snapshot-removing-request",
+                        vm=s["vm"],
+                        snapshot_name=s["snapshot"],
+                        snapshot_key=key,
+                    )
+                    del consul.kv[key]
