--- src/backy/rbd/sources/flyingcircus/__init__.py
@@ -1,1 +1,0 @@
-# Make this a package.
